refactored PostList into component and placed those into author and home pages
