+++ front-end/src/pages/HomePage.tsx
@@ -2,9 +2,7 @@ import axios from 'axios';
 import React, { useEffect, useState } from 'react';
 import { RouteComponentProps } from 'react-router-dom';
 import { Row, Col, Form, Input, Button } from 'reactstrap';
-import CreateEditPostModal from '../components/CreateEditPostModal';
-import DeletePostModal from '../components/DeleteModal';
-import PostListItem from '../components/PostListItem';
+import PostList from '../components/PostList';
 import { Post } from '../types/Post';
 
 
@@ -12,10 +10,7 @@ import { Post } from '../types/Post';
 export default function HomePage(props:any) {
 
   const [userSearch,setUserSearch] = useState('');
-
   const [postEntries, setPostEntries] = useState<Post[]|undefined>(undefined);
-
-  const [isCreatePostModalOpen, setIsCreatePostModalOpen] = useState<boolean>(false);
 
   useEffect(()=>{
     axios.get(process.env.REACT_APP_API_URL + "/api/public-posts/",).then(res => {
@@ -40,45 +35,6 @@ export default function HomePage(props:any) {
     props.history.push('/authors/'+userSearch);
   }
 
-  function prependToFeed(post:Post){
-    if(postEntries !== undefined){
-      setPostEntries([post, ...postEntries])
-    }
-  }
-
-  function removeFromFeed(postID: string) {
-    if(postEntries !== undefined){
-      setPostEntries(postEntries.filter((p) => { return postID !== p.id}))
-    }
-  }
-
-  function CreatePostModal(){
-    return(
-      <React.Fragment>
-        <Button onClick={()=>setIsCreatePostModalOpen(true)}>Create Post</Button>
-        <CreateEditPostModal
-          toggle={()=>setIsCreatePostModalOpen(!isCreatePostModalOpen)}
-          isModalOpen={isCreatePostModalOpen}
-          loggedInUser={props.loggedInUser}
-          prependToFeed={prependToFeed}
-        />
-      </React.Fragment>
-    )
-  }
-    
-
-  let postListElToDisplay;
-  if (postEntries === undefined){
-    postListElToDisplay = <p>Loading...</p>;
-  } else if (postEntries.length === 0) {
-    postListElToDisplay = <p>No Entries Found</p>;
-  } else {
-    postListElToDisplay = postEntries.map((post:Post)=>
-      <PostListItem post={post} key={post.id} isEditable={true} isDeletable={true} loggedInUser={props.loggedInUser} removeFromFeed={removeFromFeed}/>
-    );
-  }
-  console.log(postListElToDisplay)
-
   return (
     <Row>
       <Col>
@@ -86,13 +42,7 @@ export default function HomePage(props:any) {
           <Input type="text" name="Author Search" placeholder="Search Authors" onChange={e => onUserSearchChange(e)}/>
         </Form>
       </Col>
-      <Col>
-        <h3>Feed</h3>
-        {postListElToDisplay}
-      </Col>
-      <Col>
-        {props.loggedInUser ? CreatePostModal() : null}
-      </Col>
+      {<PostList postEntries={postEntries} setPostEntries={setPostEntries} loggedInUser={props.loggedInUser} />}
     </Row>
   );
 }
